Updated business logic
Added business logic to create the trips.csv output file based on the input file provided.

+++ src/main/java/com/littlepay/faresystem/model/Trip.java
@@ -2,16 +2,17 @@ package com.littlepay.faresystem.model;
 
 import lombok.Data;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 @Data
 public class Trip {
     private LocalDateTime started;
     private LocalDateTime finished;
-    private int durationSecs;
+    private long durationSecs;
     private String fromStopId;
     private String toStopId;
-    private double chargeAmount;
+    private BigDecimal chargeAmount;
     private String companyId;
     private String busId;
     private String pan;
